Fix Email Reply Webhook + Slack Notification (#2018)
* email_prospect_replied send date logic

We used GeneratedMessage.date_sent but there was a bug in the Smartlead Email Sent webhook which forgot to update date_sent. Therefor all emails sent did not have date_sent correctly populated.

We need to adjust this logic to fail-safe and fall back on when the message was created if the date_sent is not present.

* email_sent updates date sent

This fixes the email_sent logic to correctly update the date_sent field on the ProspectEmail object as well as the date_sent fields and send status on the GeneratedMessage objects. Before there was no logic.

* Add Active_convo_referral mapping

We didn't have this mapping before, but we do need it

* Fixup some syntax issues with email_prospect_replied

* Formatting